+++ src/resdk/tables/__init__.py
@@ -18,11 +18,13 @@ Table classes
 """
 from .methylation import MethylationTables  # noqa
 from .microarray import MATables  # noqa
+from .ml_ready import MLTables  # noqa
 from .rna import RNATables  # noqa
 from .variant import VariantTables  # noqa
 
 __all__ = (
     "MATables",
+    "MLTables",
     "MethylationTables",
     "RNATables",
     "VariantTables",