+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 - Support upload directly to S3 bucket
 - Add support for ``Python3.9``, ``Python3.10`` and drop support for
   ``Python3.6``
+- Add ``resdk.tables.VariantTables`` class to handle variant data
 
 Fixed
 -----